Move Helm wrapper into a separate crate
