+++ rust/helm-sys/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "helm-sys"
+version = "0.1.0"
+edition = "2021"
+publish = false
+links = "helm"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+
+[build-dependencies]
+gobuild = "0.1.0-alpha.1"
